+++ docs/changelog.md
@@ -37,6 +37,7 @@ types of changes:
 -----
 
 - <span class='badge'>Added</span> Adding support for “ghost points” to measure glyph margins.
+- <span class='badge'>Added</span> Adding kerning preview to VarFont Assistant.
 
 
 0.1.2